Add questions goal progress bar on home page
