+++ src/App.jsx
@@ -59,6 +59,7 @@ import {
 import EducationalModal from "./components/LearnModal/EducationalModal";
 import SettingsMenu from "./components/SettingsMenu/SettingsMenu";
 import ThemeMenu from "./components/ThemeMenu";
+import WaveBar from "./components/WaveBar";
 
 import {
   createUser,
@@ -3945,6 +3946,11 @@ const Home = ({
 
   const [questionsAnswered, setQuestionsAnswered] =
     useState(BASE_QUESTION_COUNT);
+  const QUESTION_GOAL = 5000;
+  const questionProgress = Math.min(
+    (questionsAnswered / QUESTION_GOAL) * 100,
+    100
+  );
 
   useEffect(() => {
     const unsubscribe = subscribeToQuestionsAnswered(setQuestionsAnswered);
@@ -4429,6 +4435,15 @@ const Home = ({
                   {questionsAnswered}
                 </Text>
 
+                <Box w="80%" maxW="400px" mt={4}>
+                  <WaveBar
+                    value={questionProgress}
+                    start="#0345fc"
+                    end="#03f4fc"
+                    height={20}
+                  />
+                </Box>
+
                 <VStack>
                   <Text
                     color="white"
@@ -4493,6 +4508,14 @@ const Home = ({
                   >
                     {questionsAnswered}
                   </Text>
+                  <Box w="80%" maxW="400px">
+                    <WaveBar
+                      value={questionProgress}
+                      start="#0345fc"
+                      end="#03f4fc"
+                      height={20}
+                    />
+                  </Box>
                 </VStack>
 
                 <VStack spacing={3} align="center" mt={12}>
